perf: Refactor Login With Oauth
