+++ package.json
@@ -94,3 +94,4 @@
         "zod": "^3.23.8"
     }
 }
+
